data_reconciliation_example/pyproject.toml
[project]
name = "data_reconciliation_example"
requires-python = ">=3.9,<3.13"
version = "0.1.0"
dependencies = [
    "dagster==1.11.0",
    "dagster-dbt>=0.27.0",
    "dagster-duckdb>=0.27.0",
    "dbt-duckdb>=1.9.4",
    "pandas>=2.3.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli[local]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "data_reconciliation_example"
registry_modules = [
    "data_reconciliation_example.components.*",
]


